Struct lyon_path::default::Path

pub struct Path { /* fields omitted */ }

A simple path data structure.

It can be created using a Builder, and can be iterated over.

Methods

impl Path

pub fn builder() -> Builder

Creates a Builder to create a path.

pub fn new() -> Path

pub fn with_capacity(cap: usize) -> Path

pub fn as_slice(&self) -> PathSlice

Important traits for Iter<'l>
impl<'l> Iterator for Iter<'l> type Item = PathEvent;

pub fn iter(&self) -> Iter

Important traits for PathIter<Iter>
impl<E, Iter> Iterator for PathIter<Iter> where
E: Into<PathEvent>,
Iter: Iterator<Item = E>, type Item = PathEvent;

pub fn path_iter(&self) -> PathIter<Iter>

Important traits for &'a mut [u8]
impl<'a> Write for &'a mut [u8]impl<'a> Read for &'a [u8]

pub fn points(&self) -> &[Point]

Important traits for &'a mut [u8]
impl<'a> Write for &'a mut [u8]impl<'a> Read for &'a [u8]

pub fn mut_points(&mut self) -> &mut [Point]

Important traits for &'a mut [u8]
impl<'a> Write for &'a mut [u8]impl<'a> Read for &'a [u8]

pub fn verbs(&self) -> &[Verb]

pub fn merge(self, other: Self) -> Self

Consumes two paths and builds one that contains them.