lyon_path

Module path_buffer

A container to store multiple paths contiguously.

Structs

Builder

A Builder that appends a path to an existing PathBuffer.

BuilderWithAttributes

A Builder that appends a path to an existing PathBuffer, with custom attributes.

Iter

Iterator over the paths in a PathBufferSlice.

PathBuffer

An object that stores multiple paths contiguously.

PathBufferSlice

A view on a PathBuffer.