Struct lyon_path::PathCommands

pub struct PathCommands { /* private fields */ }

The commands of a path encoded in a single array using IDs to refer to endpoints and control points externally.

PathCommands is a good fit when the a custom endpoint and control point types are needed or when their the user needs full control over their storage.

Representation

Path commands contains a single array of 32 bits integer values encoding path commands, endpoint IDs or control point IDs.

 __________________________________________________________________________
|       |          |      |          |         |              |          |
| Begin |EndpointID| Line |EndpointID|Quadratic|ControlPointId|EndpointID| ...
|_______|__________|______|__________|_________|______________|__________|_

Example

use lyon_path::{EndpointId, PathCommands};

let mut cmds = PathCommands::builder();

cmds.begin(EndpointId(0));
cmds.line_to(EndpointId(1));
cmds.line_to(EndpointId(2));
cmds.end(true);

let cmds = cmds.build();

Implementations

impl PathCommands

pub fn builder() -> PathCommandsBuilder

Creates a PathCommandsBuilder to create path commands.

pub fn iter(&self) -> Iter<'_>

Returns an iterator over the path commands.

pub fn as_slice(&self) -> PathCommandsSlice<'_>

Returns a view on the path commands.

pub fn path_slice<'l, Endpoint, ControlPoint>( &'l self, endpoints: &'l [Endpoint], control_points: &'l [ControlPoint], ) -> CommandsPathSlice<'_, Endpoint, ControlPoint>

Returns a view on a path made of these commands with endpoint and control point slices.

pub fn events<'l, Endpoint, ControlPoint>( &'l self, endpoints: &'l [Endpoint], control_points: &'l [ControlPoint], ) -> Events<'_, Endpoint, ControlPoint>

Returns an iterator over the path, with endpoints and control points.

pub fn event(&self, id: EventId) -> IdEvent

Returns the event for a given event ID.

pub fn next_event_id_in_path(&self, id: EventId) -> Option<EventId>

Returns the next event id within the path.

pub fn next_event_id_in_sub_path(&self, id: EventId) -> EventId

Returns the next event id within the sub-path.

Loops back to the first event after the end of the sub-path.