Struct lyon_path::PathState

pub struct PathState { /* fields omitted */ }

Represents the current state of a path while it is being built.

Methods

impl PathState

pub fn new() -> Self

pub fn current_position(&self) -> Point

The current position.

pub fn start_position(&self) -> Point

The position at the start of the current sub-path.

pub fn move_to(&mut self, to: Point)

pub fn line_to(&mut self, to: Point)

pub fn close(&mut self)

pub fn quadratic_bezier_to(&mut self, ctrl: Point, to: Point)

pub fn cubic_bezier_to(&mut self, _ctrl1: Point, ctrl2: Point, to: Point)

pub fn arc(
&mut self,
center: Point,
radii: Vector,
sweep_angle: Angle,
x_rotation: Angle
)

pub fn get_smooth_cubic_ctrl(&self) -> Point

pub fn get_smooth_quadratic_ctrl(&self) -> Point

pub fn relative_to_absolute(&self, v: Vector) -> Point