Trait lyon_path::PositionStore

pub trait PositionStore {
    // Required methods
    fn get_endpoint(&self, id: EndpointId) -> Point;
    fn get_control_point(&self, id: ControlPointId) -> Point;
}

Interface for objects storing endpoints and control points positions.

This interface can be implemented by path objects themselves or via external data structures.

Required Methods

fn get_endpoint(&self, id: EndpointId) -> Point

fn get_control_point(&self, id: ControlPointId) -> Point