Trait lyon_path::builder::Build

pub trait Build {
    type PathType;
    fn build(self) -> Self::PathType;

    fn build_and_reset(&mut self) -> Self::PathType;
}

Associated Types

type PathType

The type of object that is created by this builder.

Loading content...
Required methods

fn build(self) -> Self::PathType

Builds a path object and resets the builder so that it can be used again.

fn build_and_reset(&mut self) -> Self::PathType

Builds a path object and resets the builder so that it can be used again.

Loading content...